requirements.txt
# Database dependencies
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
Flask-SQLAlchemy==3.1.1

# Environment management
python-dotenv==1.1.1

# Web framework
flask==3.1.1

# WSGI HTTP Server for production
gunicorn==21.2.0

# NLP dependencies
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Data processing
numpy==2.3.1

# HTTP requests
requests==2.32.4

# Data validation
pydantic==2.11.7

# LangChain with free LLMs
langchain-community>=0.0.10
langchain-core>=0.1.0
transformers>=4.30.0
torch>=2.0.0

# CLI and terminal output
typer==0.16.0
rich==14.0.0

# Progress bars
tqdm==4.67.1
